+++ truffle/CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Java Interop now wraps exceptions thrown by Java method invocations in host exceptions.
 * Added [JavaInterop.isHostException](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/interop/java/JavaInterop.html#isHostException-java.lang.Throwable-) and [JavaInterop.asHostException](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/interop/java/JavaInterop.html#asHostException-java.lang.Throwable-) to identify and unwrap host exceptions, respectively.
 * The profiler infrastructure (`CPUSampler`, `CPUTracer` and `MemoryTracer`) moved to a new tools suite.
+* Added [LanguageInfo.isInternal](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/nodes/LanguageInfo.html#isInternal--)
 
 ## Version 0.30